requirements.txt
Django==1.11.3
gunicorn==19.7.1
django-filter==1.1.0
djangorestframework==3.7.7
